feat(ui): Move Save Now button to page header next to world name
✅ UI improvement implemented as requested:
- Move SaveIndicator from character summary area to page header
- Position Save Now button in upper right next to 'Playing in [World Name]'
- Use PageLayout actions prop for proper header alignment
- Remove auto-save indicator from ActiveGameSession component
- Clean up unused imports and auto-save triggers

🎨 Visual improvements:
- Save Now button now prominent and easily accessible
- Better visual hierarchy with save controls at page level
- Maintains compact styling for header integration
- Cleaner character summary section without save clutter

🔧 Technical changes:
- Leverage existing PageLayout actions infrastructure
- Move auto-save hook to page level for better separation of concerns
- Maintain all existing save functionality and error handling
- Preserve responsive design and accessibility features

Addresses user feedback from screenshot showing desired button placement
in upper right area aligned with world name header.
