+++ src/app/play/page.tsx
@@ -9,6 +9,8 @@ import GameSession from '@/components/GameSession/GameSession';
 import { LoadingPulse } from '@/components/ui/LoadingState';
 import { SectionError } from '@/components/ui/ErrorDisplay';
 import { PageLayout } from '@/components/shared/PageLayout';
+import { useAutoSave } from '@/hooks/useAutoSave';
+import { SaveIndicator } from '@/components/ui/SaveIndicator';
 
 export default function PlayPage() {
   const router = useRouter();
@@ -20,6 +22,9 @@ export default function PlayPage() {
   const currentCharacterId = useCharacterStore(state => state.currentCharacterId);
   const initializeSession = useSessionStore(state => state.initializeSession);
   const currentSessionId = useSessionStore(state => state.id);
+  
+  // Auto-save functionality for header save button
+  const autoSave = useAutoSave();
   useEffect(() => {
     const setupSession = async () => {
       try {
@@ -99,8 +104,23 @@ export default function PlayPage() {
   const pageTitle = currentWorld ? `Playing in ${currentWorld.name}` : 'Game Session';
   const pageDescription = currentWorld?.genre;
 
+  // Create SaveIndicator as page action
+  const saveAction = (
+    <SaveIndicator
+      status={autoSave.status}
+      lastSaveTime={autoSave.lastSaveTime}
+      errorMessage={autoSave.errorMessage}
+      totalSaves={autoSave.totalSaves}
+      onManualSave={autoSave.triggerSave}
+      onRetryError={autoSave.retry}
+      retryable={true}
+      compact={true}
+      className="text-sm"
+    />
+  );
+
   return (
-    <PageLayout title={pageTitle} description={pageDescription} maxWidth="7xl" className="pb-0">
+    <PageLayout title={pageTitle} description={pageDescription} maxWidth="7xl" className="pb-0" actions={saveAction}>
       <GameSession worldId={currentWorldId!} />
     </PageLayout>
   );
